Fix UX problems sharing chats and using Nut without being logged in (#87)

+++ app/lib/persistence/db.ts
@@ -1,6 +1,11 @@
-import { getSupabase } from '~/lib/supabase/client';
+// When the user is not logged in, chats are kept in local storage.
+// Otherwise, they are kept in the database. The first time the user logs in,
+// any local chats are added to the database and then deleted.
+
+import { getSupabase, getCurrentUserId } from '~/lib/supabase/client';
 import { v4 as uuid } from 'uuid';
 import { getMessagesRepositoryId, type Message } from './message';
+import { assert } from '~/lib/replay/ReplayProtocolClient';
 
 export interface ChatContents {
   id: string;
@@ -22,7 +27,31 @@ function databaseRowToChatContents(d: any): ChatContents {
   };
 }
 
+const localStorageKey = 'nut-chats';
+
+function getLocalChats(): ChatContents[] {
+  const chatJSON = localStorage.getItem(localStorageKey);
+  if (!chatJSON) {
+    return [];
+  }
+  return JSON.parse(chatJSON);
+}
+
+function setLocalChats(chats: ChatContents[] | undefined): void {
+  if (chats) {
+    localStorage.setItem(localStorageKey, JSON.stringify(chats));
+  } else {
+    localStorage.removeItem(localStorageKey);
+  }
+}
+
 export async function getAllChats(): Promise<ChatContents[]> {
+  const userId = await getCurrentUserId();
+
+  if (!userId) {
+    return getLocalChats();
+  }
+
   const { data, error } = await getSupabase().from('chats').select('*');
 
   if (error) {
@@ -32,12 +61,33 @@ export async function getAllChats(): Promise<ChatContents[]> {
   return data.map(databaseRowToChatContents);
 }
 
+export async function syncLocalChats(): Promise<void> {
+  const userId = await getCurrentUserId();
+  const localChats = getLocalChats();
+
+  if (userId && localChats.length) {
+    for (const chat of localChats) {
+      await setChatContents(chat.id, chat.title, chat.messages);
+    }
+    setLocalChats(undefined);
+  }
+}
+
 export async function setChatContents(id: string, title: string, messages: Message[]): Promise<void> {
-  const { data: user } = await getSupabase().auth.getUser();
-  const userId = user.user?.id;
+  const userId = await getCurrentUserId();
 
   if (!userId) {
-    throw new Error('Not logged in');
+    const localChats = getLocalChats().filter((c) => c.id != id);
+    localChats.push({
+      id,
+      title,
+      messages,
+      repositoryId: getMessagesRepositoryId(messages),
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+    });
+    setLocalChats(localChats);
+    return;
   }
 
   const repositoryId = getMessagesRepositoryId(messages);
@@ -55,7 +105,30 @@ export async function setChatContents(id: string, title: string, messages: Messa
   }
 }
 
-export async function getChatContents(id: string): Promise<ChatContents> {
+export async function getChatPublicData(id: string): Promise<{ repositoryId: string; title: string }> {
+  const { data, error } = await getSupabase().rpc('get_chat_public_data', { chat_id: id });
+
+  if (error) {
+    throw error;
+  }
+
+  if (data.length != 1) {
+    throw new Error(`Unknown chat ${id}`);
+  }
+
+  return {
+    repositoryId: data[0].repository_id,
+    title: data[0].title,
+  };
+}
+
+export async function getChatContents(id: string): Promise<ChatContents | undefined> {
+  const userId = await getCurrentUserId();
+
+  if (!userId) {
+    return getLocalChats().find((c) => c.id == id);
+  }
+
   const { data, error } = await getSupabase().from('chats').select('*').eq('id', id);
 
   if (error) {
@@ -63,13 +136,21 @@ export async function getChatContents(id: string): Promise<ChatContents> {
   }
 
   if (data.length != 1) {
-    throw new Error('Unexpected chat contents returned');
+    return undefined;
   }
 
   return databaseRowToChatContents(data[0]);
 }
 
 export async function deleteById(id: string): Promise<void> {
+  const userId = await getCurrentUserId();
+
+  if (!userId) {
+    const localChats = getLocalChats().filter((c) => c.id != id);
+    setLocalChats(localChats);
+    return;
+  }
+
   const { error } = await getSupabase().from('chats').delete().eq('id', id);
 
   if (error) {
@@ -83,8 +164,9 @@ export async function createChat(title: string, messages: Message[]): Promise<st
   return id;
 }
 
-export async function updateChatTitle(id: string, title: string): Promise<void> {
+export async function databaseUpdateChatTitle(id: string, title: string): Promise<void> {
   const chat = await getChatContents(id);
+  assert(chat, 'Unknown chat');
 
   if (!title.trim()) {
     throw new Error('Title cannot be empty');
